Renaming context "switch" to "set"
