+++ cli/cli/commands/kurtosis_context/kurtosis_context.go
@@ -3,9 +3,9 @@ package kurtosis_context
 import (
 	"github.com/kurtosis-tech/kurtosis/cli/cli/command_str_consts"
 	"github.com/kurtosis-tech/kurtosis/cli/cli/commands/kurtosis_context/add"
-	"github.com/kurtosis-tech/kurtosis/cli/cli/commands/kurtosis_context/context_switch"
 	"github.com/kurtosis-tech/kurtosis/cli/cli/commands/kurtosis_context/ls"
 	"github.com/kurtosis-tech/kurtosis/cli/cli/commands/kurtosis_context/rm"
+	"github.com/kurtosis-tech/kurtosis/cli/cli/commands/kurtosis_context/set"
 	"github.com/spf13/cobra"
 )
 
@@ -21,5 +21,5 @@ func init() {
 	ContextCmd.AddCommand(add.ContextAddCmd.MustGetCobraCommand())
 	ContextCmd.AddCommand(ls.ContextLsCmd.MustGetCobraCommand())
 	ContextCmd.AddCommand(rm.ContextRmCmd.MustGetCobraCommand())
-	ContextCmd.AddCommand(context_switch.ContextSwitchCmd.MustGetCobraCommand())
+	ContextCmd.AddCommand(set.ContextSetCmd.MustGetCobraCommand())
 }
